refactor server setup
