+++ binaries/scheduler/main.go
@@ -2,99 +2,16 @@ package main
 
 import (
 	"flag"
-	"fmt"
-	"log"
-
-	"github.com/apache/thrift/lib/go/thrift"
-	"github.com/scootdev/scoot/config/jsonconfig"
-	"github.com/scootdev/scoot/ice"
-
-	// For putting into ice.MagicBag
-	"github.com/scootdev/scoot/common/endpoints"
-	"github.com/scootdev/scoot/common/stats"
-	"github.com/scootdev/scoot/saga"
-	"github.com/scootdev/scoot/sched/scheduler"
 	"github.com/scootdev/scoot/scootapi/server"
-
-	// For putting into jsonconfig.Options
-	"github.com/scootdev/scoot/config/scootconfig"
-	"github.com/scootdev/scoot/saga/sagalogs"
 )
 
-var addr = flag.String("addr", "localhost:9090", "Bind address for api server.")
-var httpPort = flag.Int("http_port", 9091, "port to serve http on")
-var cfgText = flag.String("sched_config", "", "Scheduler Configuration.")
-
-type servers struct {
-	thrift thrift.TServer
-	http   *endpoints.TwitterServer
-}
-
-func makeServers(
-	thrift thrift.TServer,
-	http *endpoints.TwitterServer) servers {
-	return servers{thrift, http}
-}
-
 func main() {
-	log.Println("Starting Cloud Scoot API Server & Scheduler")
-	flag.Parse()
-
-	bag := ice.NewMagicBag()
-	bag.PutMany(
-		func() (thrift.TServerTransport, error) { return thrift.NewTServerSocket(*addr) },
-		endpoints.MakeStatsReceiver,
-		server.MakeServer,
-		server.NewHandler,
-		func(s stats.StatsReceiver) *endpoints.TwitterServer {
-			return endpoints.NewTwitterServer(fmt.Sprintf(":%d", *httpPort), s)
-		},
-		makeServers,
-		saga.MakeSagaCoordinator,
-		sagalogs.MakeInMemorySagaLog,
-		scheduler.NewStatefulSchedulerFromCluster,
-		thrift.NewTTransportFactory,
-		func() thrift.TProtocolFactory {
-			return thrift.NewTBinaryProtocolFactoryDefault()
-		},
-	)
-
-	schema := jsonconfig.Schema(map[string]jsonconfig.Implementations{
+	config := `{
 		"Cluster": {
-			"memory": &scootconfig.ClusterMemoryConfig{},
-			"local":  &scootconfig.ClusterLocalConfig{},
-			"": &scootconfig.ClusterMemoryConfig{
-				Type:  "memory",
-				Count: 10,
-			},
-		},
-		"Workers": {
-			"local": &scootconfig.WorkersLocalConfig{},
-			"rpc":   &scootconfig.WorkersThriftConfig{},
-			"":      &scootconfig.WorkersLocalConfig{Type: "local"},
-		},
-	})
-
-	mod, err := schema.Parse([]byte(*cfgText))
-	if err != nil {
-		log.Fatal("Error configuring Scoot API: ", err)
-	}
-	bag.InstallModule(mod)
-
-	// TODO(dbentley): we may want to refactor all of the code below here to be
-	// logic in a library instead of copied into each main, but for now be explicit.
-	var servers servers
-	err = bag.Extract(&servers)
-	if err != nil {
-		log.Fatal("Error injecting servers", err)
-	}
-
-	errCh := make(chan error)
-	go func() {
-		errCh <- servers.http.Serve()
-	}()
-	go func() {
-		errCh <- servers.thrift.Serve()
-	}()
-	log.Fatal("Error serving: ", <-errCh)
+			"Type": "memory",
+			"Count": 10
+		}
+	}`
+	bag, schema := server.SetupServer()
+	server.RunServer(bag, schema, []byte(config))
 }
